Upload

+++ WorkAptech/WorkAptech/Areas/Admin/Controllers/CategoriesController.cs
@@ -1,11 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using WorkAptech.Data;
+using WorkAptech.Utility;
 
 namespace WorkAptech.Areas.Admin.Controllers
 {
@@ -13,10 +16,12 @@ namespace WorkAptech.Areas.Admin.Controllers
     public class CategoriesController : Controller
     {
         private readonly ApplicationDbContext _context;
-
-        public CategoriesController(ApplicationDbContext context)
+        //upload Img
+        private readonly IHostingEnvironment _hostingEnvironment;
+        public CategoriesController(ApplicationDbContext context, IHostingEnvironment hostingEnvironment)
         {
             _context = context;
+            _hostingEnvironment = hostingEnvironment;
         }
 
         // GET: Admin/Categories
@@ -56,13 +61,41 @@ namespace WorkAptech.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name,Picture")] Category category)
         {
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
-                _context.Add(category);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                return View(category);
             }
-            return View(category);
+            _context.Add(category);
+            await _context.SaveChangesAsync();
+
+            //Work on the image saving section
+
+            string webRootPath = _hostingEnvironment.WebRootPath;
+            var files = HttpContext.Request.Form.Files;
+
+            var CateFromDb = await _context.Category.FindAsync(category.Id);
+
+            if (files.Count > 0)
+            {
+                //files has been uploaded
+                var uploads = Path.Combine(webRootPath, "images");
+                var extension = Path.GetExtension(files[0].FileName);
+
+                using (var filesStream = new FileStream(Path.Combine(uploads, category.Id + extension), FileMode.Create))
+                {
+                    files[0].CopyTo(filesStream);
+                }
+                CateFromDb.Picture = @"\images\category" + category.Id + extension;
+            }
+            else
+            {
+                //no file was uploaded, so use default
+                var uploads = Path.Combine(webRootPath, @"images\category" + SD.DefaultCategoryImage);
+                System.IO.File.Copy(uploads, webRootPath + @"\images\category" + category.Id + ".png");
+                CateFromDb.Picture = @"\images\category" + category.Id + ".png";
+            }
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Admin/Categories/Edit/5
@@ -97,7 +130,34 @@ namespace WorkAptech.Areas.Admin.Controllers
             {
                 try
                 {
-                    _context.Update(category);
+                    string webRootPath = _hostingEnvironment.WebRootPath;
+                    var files = HttpContext.Request.Form.Files;
+
+                    var CategoryFromDb = await _context.Category.FindAsync(category.Id);
+
+                    if (files.Count > 0)
+                    {
+                        //New Image has been uploaded
+                        var uploads = Path.Combine(webRootPath, "images\\category");
+                        var extension_new = Path.GetExtension(files[0].FileName);
+
+                        //Delete the original file
+                        var imagePath = Path.Combine(webRootPath, CategoryFromDb.Picture.TrimStart('\\'));
+
+                        if (System.IO.File.Exists(imagePath))
+                        {
+                            System.IO.File.Delete(imagePath);
+                        }
+
+                        //we will upload the new file
+                        using (var filesStream = new FileStream(Path.Combine(uploads, category.Id + extension_new), FileMode.Create))
+                        {
+                            files[0].CopyTo(filesStream);
+                        }
+                        CategoryFromDb.Picture = @"\images\category" + category.Id + extension_new;
+                    }
+                    CategoryFromDb.Name = category.Name;
+                    // _context.Update(category);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
@@ -139,9 +199,18 @@ namespace WorkAptech.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            string webRootPath = _hostingEnvironment.WebRootPath;
             var category = await _context.Category.FindAsync(id);
-            _context.Category.Remove(category);
-            await _context.SaveChangesAsync();
+            if (category != null)
+            {
+                var imagePath = Path.Combine(webRootPath, category.Picture.TrimStart('\\'));
+                if (System.IO.File.Exists(imagePath))
+                {
+                    System.IO.File.Delete(imagePath);
+                }
+                _context.Category.Remove(category);
+                await _context.SaveChangesAsync();
+            }
             return RedirectToAction(nameof(Index));
         }
